pyproject.toml
[project]
name = "finetune-workshop"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "ffmpeg-python>=0.2.0",
    "opencv-python>=4.12.0.88",
    "ultralytics>=8.3.235",
]
